refactor: ReservationController에서 상태로 ReservationService 가지도록 변경
